chore: cleaned up cli code feat: migration script would run whenever app tries to ues the tortorise orm
