+++ README.md
@@ -1,22 +1,26 @@
-# Handshake 🤝
+# Handshake 🫱🏾‍🫲🏼
 _A Modern Test Reporter_
 
 Processes the data collected from the test execution and then generates a modern Dashboard.
 
 ## Getting Started
 
-### Nodejs-Reporters
-
-#### WebdriverIO
+### WebdriverIO
 
 [WebdriverIO](https://webdriver.io/) is a Next-gen browser and mobile automation test framework for Node.js
 
-with the help of customer [reporter](https://webdriver.io/docs/customreporter/) and [service](https://webdriver.io/docs/customservices), we would be able to collect the results generated through its project.
-
 1. Install Packages: `npm install wdio-handshake-reporter`
+2. Install Dashboard: `npm install handshake`
+3. Once done you can configure the reporter in your configuration.
 
 ## Plans
+below are the things I would like to implement sooner or later.
 
-### Test Reports
-- Dynamic Report
-- JIRA Reporter
+* [X] Add support for `webdriverio`
+* [ ] Add Reporter for `jest`
+* [ ] Work on documentation
+* [ ] Support for `mocha`
+* [ ] Listen to `cypress` results
+* [ ] Add Support for `playwright`
+
+## Sample Shots
