+++ handshake/services/SchedularService/register.py
@@ -2,10 +2,13 @@ from loguru import logger
 from handshake.services.DBService.models.dynamic_base import TaskBase, JobType
 from uuid import uuid4, UUID
 from typing import Union
-from handshake.services.DBService.models.static_base import (
-    AttachmentType,
-)
 from handshake.services.DBService.models.result_base import TestLogBase, LogType
+from handshake.services.DBService.lifecycle import init_tortoise_orm, close_connection
+from handshake.services.DBService.models.config_base import ExportBase
+from handshake.services.DBService.models.result_base import RunBase
+from pathlib import Path
+from typing import List
+from os import getenv
 
 
 async def register_patch_suite(suiteID: str, testID: str) -> TaskBase:
@@ -42,3 +45,28 @@ async def warn_about_test_run(test_id: Union[str, UUID], about: str, **extra) ->
         test_id=str(test_id), message=about, type=LogType.WARN, feed=extra
     )
     return True
+
+
+async def createExportTicket(
+    maxTestRuns: int, path: Path, store: List[str], runs: List[str], clarity
+):
+    await init_tortoise_orm(path)
+    ticket = await ExportBase.create(
+        maxTestRuns=maxTestRuns, clarity="" if not clarity else getenv("CLARITY")
+    )
+
+    runs.extend(
+        await RunBase.all()
+        .order_by("-started")
+        .limit(maxTestRuns)
+        .values_list("testID", flat=True)
+    )
+
+    await close_connection()
+    store.append(str(ticket.ticketID))
+
+
+async def deleteExportTicket(ticketID: str):
+    task = await ExportBase.filter(ticketID=ticketID).first()
+    logger.info("Deleting the Export ticket: {}", ticketID)
+    await task.delete()
